ai-microservices/Face-match/requirements.txt
flask
face_recognition
numpy
pillow
dlib==19.24.0